Cleanup Code and downgrade dependencies to reliable versions
