+++ components/paragraph/stanford-banner.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 
-import {BannerParagraph, DrupalImageMedia, DrupalLink} from "@/lib/drupal/drupal";
-import Banner from "../patterns/banner";
+import {DrupalImageMedia, DrupalLink} from "@/lib/drupal/drupal";
+import Banner from "@/components/patterns/banner";
 import {PropsWithoutRef} from "react";
 
 interface Props extends PropsWithoutRef<any> {
