+++ components/paragraph/stanford-media-caption.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 
 import {DrupalImageMedia, DrupalLink} from "@/lib/drupal/drupal";
 import formatHtml from "@/lib/format-html";
-import Oembed from "../patterns/oembed";
+import Oembed from "@/components/patterns/oembed";
 import Link from "next/link";
 import {PropsWithoutRef} from "react";
